refactor: modularize LunaTask API client to enforce 500-line limit
Split monolithic client.py (784 lines) into focused modules:
- BaseClient: HTTP infrastructure, auth, retries, rate limiting
- TasksClientMixin: Task CRUD operations
- NotesClientMixin: Note creation
- JournalClientMixin: Journal entry creation
- HabitsClientMixin: Habit tracking
- BaseClientProtocol: Typing protocol for mixins

Key improvements:
- Reduced client.py from 784 to 46 lines
- Added comprehensive characterization tests for pre-refactor behavior
- Centralized rate limiting and mutation delay in base client
- Fixed create_journal_entry implementation and JournalEntryResponse model configuration
- Maintained public API compatibility and test coverage ≥95%
- Enhanced type safety with strict protocols for mixin composition

All functionality preserved with identical public interface.
